# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartLib Pro - Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List books
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: authorId
          in: query
          schema:
            type: ["integer", "null"]
        - name: publisherId
          in: query
          schema:
            type: ["integer", "null"]
        - name: available
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Books list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookList"
    post:
      operationId: createBook
      tags: [books]
      summary: Create book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Authors
...
  /borrow-records:
    get:
      operationId: listBorrowRecords
      tags: [borrowing]
      summary: List borrow records
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, returned, overdue]
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: bookId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Borrow records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowRecordList"
    post:
      operationId: createBorrowRecord
      tags: [borrowing]
      summary: Borrow a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowRecord"
  /borrow-records/{id}/return:
    post:
      operationId: returnBook
      tags: [borrowing]
      summary: Return a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowRecord"
  /borrow-records/{id}/renew:
    post:
      operationId: renewBook
      tags: [borrowing]
      summary: Renew a borrow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowRecord"

  # Reservations
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # User
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, librarian, student]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        studentId:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Book:
      type: object
      required: [id, title, isbn, totalCopies, availableCopies, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        isbn:
          type: string
        description:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        publishedYear:
          type: ["integer", "null"]
        language:
          type: string
        shelfLocation:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        publisherId:
          type: ["integer", "null"]
        author:
          $ref: "#/components/schemas/Author"
        category:
          $ref: "#/components/schemas/Category"
        publisher:
          $ref: "#/components/schemas/Publisher"
        createdAt:
          type: string
          format: date-time
    BookInput:
      type: object
      required: [title, isbn, totalCopies]
      properties:
        title:
          type: string
        isbn:
          type: string
        description:
          type: string
        coverUrl:
          type: string
        totalCopies:
          type: integer
        publishedYear:
          type: integer
        language:
          type: string
        shelfLocation:
          type: string
        authorId:
          type: integer
        categoryId:
          type: integer
        publisherId:
          type: integer
    BookUpdate:
      type: object
      properties:
        title:
          type: string
        isbn:
          type: string
        description:
          type: string
        coverUrl:
          type: string
        totalCopies:
          type: integer
        publishedYear:
          type: integer
        language:
          type: string
        shelfLocation:
          type: string
        authorId:
          type: integer
        categoryId:
          type: integer
        publisherId:
          type: integer
    BookList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Author
    Author:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
...
    Publisher:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
...
    BorrowRecord:
      type: object
      required: [id, userId, bookId, borrowedAt, dueDate, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bookId:
          type: integer
        borrowedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnedAt:
          type: ["string", "null"]
          format: date-time
        renewCount:
          type: integer
        status:
          type: string
          enum: [active, returned, overdue]
        user:
          $ref: "#/components/schemas/User"
        book:
          $ref: "#/components/schemas/Book"
    BorrowInput:
      type: object
      required: [userId, bookId]
      properties:
        userId:
          type: integer
        bookId:
          type: integer
        durationDays:
          type: integer
          default: 14
    BorrowRecordList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BorrowRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Reservation
...