# lib/api-zod/src/generated/types/borrowInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SmartLib Pro - Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface BorrowInput {
  userId: number;
  bookId: number;
  durationDays?: number;
}